# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Invoice & Receipt System API
  version: 0.1.0
  description: |
    Internal API for a small Israeli business operating as `עוסק פטור`.

    Conventions that hold everywhere:

    - Money is an **integer number of agorot** (1 ILS = 100 agorot). Never a
      decimal, never a float.
    - Timestamps are RFC 3339 in UTC. Business dates are `YYYY-MM-DD`, already
      interpreted in `Asia/Jerusalem`.
    - Authentication is a session cookie. Unsafe methods must also send the
      `X-CSRF-Token` header, echoing the `csrf_token` cookie.
    - Errors always use one envelope. `code` is a stable English identifier;
      `message` is Hebrew and is shown to the user as-is.

    Sending a document never affects the document: delivery runs in a
    background job, after issuance has committed, so a failed send cannot roll
    back an issued document (plan.md 14).

    Critical commands accept an `Idempotency-Key` header. A replay returns the
    original response with `Idempotent-Replay: true`, so a retry after a timeout
    never issues a second document with a second official number.
servers:
  - url: /api/v1
...
security:
  - sessionCookie: []
paths:
  /auth/login:
    post:
      tags: [auth]
      summary: Start a session
      description: |
        Sets the `session` and `csrf_token` cookies. Every failure — wrong
        password, unknown account, inactive account — returns the same
        `INVALID_CREDENTIALS` error, so accounts cannot be enumerated.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string, format: email }
                password: { type: string, format: password }
      responses:
        '200':
          description: Session created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/MeResponse' }
        '401':
          $ref: '#/components/responses/Error'
        '422':
          $ref: '#/components/responses/Error'
        '429':
          description: Rate limited, or the account is temporarily locked
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorEnvelope' }
  /auth/logout:
    post:
      tags: [auth]
      summary: End the current session
      responses:
        '204': { description: Session ended }
        '401': { $ref: '#/components/responses/Error' }
        '403': { $ref: '#/components/responses/Error' }
  /auth/me:
    get:
      tags: [auth]
      summary: The authenticated user
      responses:
        '200':
          description: Current user
          content:
            application/json:
              schema: { $ref: '#/components/schemas/MeResponse' }
        '401': { $ref: '#/components/responses/Error' }
  /auth/change-password:
    post:
      tags: [auth]
      summary: Change your own password
      description: |
        On success every session of the user is destroyed, including the one
        making the request, so the client must log in again.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [current_password, new_password]
              properties:
                current_password: { type: string, format: password }
                new_password: { type: string, format: password, minLength: 12 }
      responses:
        '204': { description: Password changed; all sessions ended }
        '401': { $ref: '#/components/responses/Error' }
        '422': { $ref: '#/components/responses/Error' }
...
  /customers/:
    get:
      tags: [customers]
      summary: List customers
      description: |
        Substring search over name, legal name, business number, phone (as typed
        and as bare digits) and email. Archived customers are excluded unless
        `include_archived` is true.
      parameters:
        - $ref: '#/components/parameters/SearchQuery'
        - $ref: '#/components/parameters/IncludeArchived'
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
      responses:
        '200':
          description: A page of customers
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CustomerPage' }
    post:
      tags: [customers]
      summary: Create a customer
      description: |
        When the input resembles an existing customer the request is refused
        with `409` and the candidates attached, unless `confirm_duplicate` is
        true. Duplicates are warned about, never blocked outright: two people
        can legitimately share a phone number (plan.md 8).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              allOf:
                - $ref: '#/components/schemas/CustomerInput'
                - type: object
                  properties:
                    confirm_duplicate:
                      type: boolean
                      default: false
                      description: Proceed even though duplicates were found
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Customer' }
        '403': { $ref: '#/components/responses/Error' }
        '409':
          description: Possible duplicates found
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DuplicateConflict' }
        '422': { $ref: '#/components/responses/Error' }
  /customers/check-duplicates:
    post:
      tags: [customers]
      summary: Look for customers resembling this input, without writing
      description: The UI calls this while the operator types.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              allOf:
                - $ref: '#/components/schemas/CustomerInput'
                - type: object
                  properties:
                    exclude_id:
                      type: string
                      format: uuid
                      description: The customer being edited, so it never matches itself
      responses:
        '200':
          description: Candidates
          content:
            application/json:
              schema:
                type: object
                properties:
                  duplicates:
                    type: array
                    items: { $ref: '#/components/schemas/CustomerDuplicate' }
  /customers/{id}:
    parameters:
      - $ref: '#/components/parameters/ResourceId'
    get:
      tags: [customers]
      summary: One customer
      responses:
        '200':
          description: Customer
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Customer' }
        '404': { $ref: '#/components/responses/Error' }
    put:
      tags: [customers]
      summary: Replace a customer's details
      description: |
        Does not affect documents already issued: each one carries its own
        snapshot of the customer (plan.md 11).
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CustomerInput' }
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Customer' }
        '404': { $ref: '#/components/responses/Error' }
        '422': { $ref: '#/components/responses/Error' }
  /customers/{id}/active:
    parameters:
      - $ref: '#/components/parameters/ResourceId'
    put:
      tags: [customers]
      summary: Archive or restore a customer
      description: |
        There is no delete. Documents reference customers and must stay readable
        for as long as the law requires them kept.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [active]
              properties:
                active: { type: boolean }
                reason: { type: string }
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Customer' }
...
  /backups/status:
    get:
      tags: [backups]
      summary: Where backups stand (OWNER only)
      description: |
        `stale` is true when the newest successful backup is more than 36 hours
        old, or when none has ever succeeded — a daily schedule has then missed
        at least one run. `tools` reports whether `pg_dump` matches the server:
        a mismatch produces archives that do not restore cleanly, which is worse
        than a visibly failed backup.
      responses:
        '200':
          description: Status
          content:
            application/json:
              schema: { $ref: '#/components/schemas/BackupStatus' }
        '403': { $ref: '#/components/responses/Error' }
...
  /backups/run:
    post:
      tags: [backups]
      summary: Take a backup now (OWNER only)
      description: |
        Runs synchronously rather than queued: an operator asking for a backup is
        usually about to do something risky and wants to know it worked first.
        The same thing is available without a login as `api -backup-now`.
      responses:
        '200':
          description: The completed run
          content:
            application/json:
              schema: { $ref: '#/components/schemas/BackupRun' }
        '403': { $ref: '#/components/responses/Error' }
        '500':
          description: No backup directory is configured, or pg_dump failed
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorEnvelope' }
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: session
      description: |
        Set by `POST /auth/login`. Unsafe methods must also send
        `X-CSRF-Token`, echoing the `csrf_token` cookie.
  parameters:
    ResourceId:
      name: id
      in: path
      required: true
      schema: { type: string, format: uuid }
    SearchQuery:
      name: q
      in: query
      schema: { type: string }
      description: Case-insensitive substring match
    IncludeArchived:
      name: include_archived
      in: query
      schema: { type: boolean, default: false }
    Limit:
      name: limit
      in: query
      schema: { type: integer, minimum: 1, maximum: 200, default: 50 }
    Offset:
      name: offset
      in: query
      schema: { type: integer, minimum: 0, default: 0 }
...
  responses:
    Error:
      description: Error
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
  schemas:
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              description: Stable English machine identifier
              examples: [UNAUTHORIZED, VALIDATION_FAILED, DOCUMENT_IMMUTABLE]
            message:
              type: string
              description: Hebrew text, shown to the user as-is
            details:
              type: object
              additionalProperties: true
              description: Field-level messages, keyed by field name
            request_id:
              type: string
    Role:
      type: string
      enum: [OWNER, OPERATOR, ACCOUNTANT, READ_ONLY]
    User:
      type: object
      properties:
        id: { type: string, format: uuid }
        email: { type: string, format: email }
        display_name: { type: string }
        active: { type: boolean }
        must_change_password: { type: boolean }
        roles:
          type: array
          items: { $ref: '#/components/schemas/Role' }
        last_login_at: { type: [string, 'null'], format: date-time }
        created_at: { type: string, format: date-time }
    MeResponse:
      type: object
      properties:
        user: { $ref: '#/components/schemas/User' }
        roles:
          type: array
          items: { type: string }
...
    CustomerType:
      type: string
      enum: [PERSON, BUSINESS]
    DeliveryPreference:
      type: string
      enum: [EMAIL, WHATSAPP, NONE]
      description: |
        How this customer prefers to receive documents. EMAIL requires an email
        address and WHATSAPP a phone number.
    CustomerInput:
      type: object
      required: [display_name]
      properties:
        customer_type:
          allOf: [{ $ref: '#/components/schemas/CustomerType' }]
          default: PERSON
        display_name: { type: string, minLength: 1 }
        legal_name: { type: string }
        business_or_id_number:
          type: string
          description: Company number or national ID. Optional by design (plan.md 8).
        phone: { type: string }
        email: { type: string }
        address: { type: string }
        notes: { type: string }
        preferred_delivery:
          allOf: [{ $ref: '#/components/schemas/DeliveryPreference' }]
          default: NONE
    Customer:
      allOf:
        - $ref: '#/components/schemas/CustomerInput'
        - type: object
          properties:
            id: { type: string, format: uuid, readOnly: true }
            active: { type: boolean, readOnly: true }
            created_at: { type: string, format: date-time, readOnly: true }
            updated_at: { type: string, format: date-time, readOnly: true }
    CustomerPage:
      type: object
      properties:
        customers:
          type: array
          items: { $ref: '#/components/schemas/Customer' }
        total: { type: integer }
        limit: { type: integer }
        offset: { type: integer }
    CustomerDuplicate:
      type: object
      properties:
        customer: { $ref: '#/components/schemas/Customer' }
        reason:
          type: string
          enum: [PHONE, EMAIL, BUSINESS_NUMBER, NAME]
        reason_hebrew: { type: string }
    DuplicateConflict:
      type: object
      properties:
        error:
          type: object
          properties:
            code: { type: string, examples: [CONFLICT] }
            message: { type: string }
            details:
              type: object
              properties:
                code: { type: string, examples: [DUPLICATE_CUSTOMER] }
                duplicates:
                  type: array
                  items: { $ref: '#/components/schemas/CustomerDuplicate' }
...
    BackupRun:
      type: object
      properties:
        id: { type: string, format: uuid }
        started_at: { type: string, format: date-time }
        finished_at: { type: [string, 'null'], format: date-time }
        state:
          type: string
          enum: [RUNNING, SUCCEEDED, FAILED]
        archive_path: { type: [string, 'null'] }
        archive_bytes: { type: [integer, 'null'], format: int64 }
        archive_sha256: { type: [string, 'null'] }
        file_count: { type: [integer, 'null'] }
        trigger:
          type: string
          enum: [SCHEDULED, MANUAL]
        error: { type: [string, 'null'] }
    BackupStatus:
      type: object
      properties:
        configured:
          type: boolean
          description: False when BACKUP_DIR is unset; no backups run at all
        directory: { type: string }
        retention_days:
          type: integer
          description: |
            How many archives stay in the folder. Disk hygiene, not records
            retention — financial records are never pruned.
        schedule_local: { type: string }
        last_succeeded_at: { type: [string, 'null'], format: date-time }
        last_archive_bytes: { type: [integer, 'null'], format: int64 }
        hours_since_success: { type: [number, 'null'] }
        stale:
          type: boolean
          description: The newest good backup is over 36 hours old, or there is none
        last_run_state: { type: [string, 'null'] }
        last_run_error: { type: [string, 'null'] }
        tools:
          type: object
          properties:
            pg_dump_major: { type: integer }
            server_major: { type: integer }
            version_mismatch: { type: boolean }
            version_warning: { type: string }